--- a/src/constants.js
+++ b/src/constants.js
@@ -123,3 +123,6 @@ export var RGBM16Encoding = 3005;
 export var RGBDEncoding = 3006;
 export var BasicDepthPacking = 3200;
 export var RGBADepthPacking = 3201;
+export var AutoInstancingDisabled = 0;
+export var AutoInstancingAttributes = 1;
+export var AutoInstancingTexture = 2;