--- a/src/renderers/shaders/ShaderChunk/common_vertex.glsl
+++ b/src/renderers/shaders/ShaderChunk/common_vertex.glsl
@@ -1,0 +1,7 @@
+#ifdef USE_INSTANCING
+
+    mat4 modelMatrix = getModelMatrix();
+    mat4 modelViewMatrix = viewMatrix * modelMatrix;
+    mat3 normalMatrix = getNormalMatrix();
+
+#endif