--- a/src/renderers/shaders/ShaderLib/depth_vert.glsl
+++ b/src/renderers/shaders/ShaderLib/depth_vert.glsl
@@ -8,6 +8,7 @@
 
 void main() {
 
+	#include <common_vertex>
 	#include <uv_vertex>
 
 	#include <skinbase_vertex>