--- a/src/renderers/shaders/ShaderLib/linedashed_vert.glsl
+++ b/src/renderers/shaders/ShaderLib/linedashed_vert.glsl
@@ -11,6 +11,7 @@ varying float vLineDistance;
 
 void main() {
 
+	#include <common_vertex>
 	#include <color_vertex>
 
 	vLineDistance = scale * lineDistance;