--- a/src/renderers/shaders/ShaderLib/meshbasic_vert.glsl
+++ b/src/renderers/shaders/ShaderLib/meshbasic_vert.glsl
@@ -11,6 +11,7 @@
 
 void main() {
 
+	#include <common_vertex>
 	#include <uv_vertex>
 	#include <uv2_vertex>
 	#include <color_vertex>