--- a/src/renderers/shaders/ShaderLib/meshlambert_vert.glsl
+++ b/src/renderers/shaders/ShaderLib/meshlambert_vert.glsl
@@ -24,6 +24,7 @@ varying vec3 vLightFront;
 
 void main() {
 
+	#include <common_vertex>
 	#include <uv_vertex>
 	#include <uv2_vertex>
 	#include <color_vertex>