--- a/src/renderers/shaders/ShaderLib/meshphysical_vert.glsl
+++ b/src/renderers/shaders/ShaderLib/meshphysical_vert.glsl
@@ -23,6 +23,7 @@ varying vec3 vViewPosition;
 
 void main() {
 
+	#include <common_vertex>
 	#include <uv_vertex>
 	#include <uv2_vertex>
 	#include <color_vertex>