--- a/src/renderers/shaders/ShaderLib/shadow_vert.glsl
+++ b/src/renderers/shaders/ShaderLib/shadow_vert.glsl
@@ -2,6 +2,7 @@
 
 void main() {
 
+	#include <common_vertex>
 	#include <begin_vertex>
 	#include <project_vertex>
 	#include <worldpos_vertex>