--- a/src/renderers/webgl/WebGLUniforms.js
+++ b/src/renderers/webgl/WebGLUniforms.js
@@ -49,6 +49,7 @@
 
 import { CubeTexture } from '../../textures/CubeTexture';
 import { Texture } from '../../textures/Texture';
+import { _Math } from '../../math/Math';
 
 var emptyTexture = new Texture();
 var emptyCubeTexture = new CubeTexture();
@@ -109,7 +110,7 @@ function flatten( array, nBlocks, blockSize ) {
 
 // Texture unit allocation
 
-function allocTexUnits( renderer, n ) {
+function allocTexUnits( renderer, uuid, n ) {
 
 	var r = arrayCacheI32[ n ];
 
@@ -121,7 +122,7 @@ function allocTexUnits( renderer, n ) {
 	}
 
 	for ( var i = 0; i !== n; ++ i )
-		r[ i ] = renderer.allocTextureUnit();
+		r[ i ] = renderer.allocTextureUnit( uuid + n );
 
 	return r;
 
@@ -188,7 +189,7 @@ function setValue4fm( gl, v ) {
 
 function setValueT1( gl, v, renderer ) {
 
-	var unit = renderer.allocTextureUnit();
+	var unit = renderer.allocTextureUnit( this.uuid );
 	gl.uniform1i( this.addr, unit );
 	renderer.setTexture2D( v || emptyTexture, unit );
 
@@ -196,7 +197,7 @@ function setValueT1( gl, v, renderer ) {
 
 function setValueT6( gl, v, renderer ) {
 
-	var unit = renderer.allocTextureUnit();
+	var unit = renderer.allocTextureUnit( this.uuid );
 	gl.uniform1i( this.addr, unit );
 	renderer.setTextureCube( v || emptyCubeTexture, unit );
 
@@ -285,7 +286,7 @@ function setValueM4a( gl, v ) {
 function setValueT1a( gl, v, renderer ) {
 
 	var n = v.length,
-		units = allocTexUnits( renderer, n );
+		units = allocTexUnits( renderer, this.uuid, n );
 
 	gl.uniform1iv( this.addr, units );
 
@@ -300,7 +301,7 @@ function setValueT1a( gl, v, renderer ) {
 function setValueT6a( gl, v, renderer ) {
 
 	var n = v.length,
-		units = allocTexUnits( renderer, n );
+		units = allocTexUnits( renderer, this.uuid, n );
 
 	gl.uniform1iv( this.addr, units );
 
@@ -343,6 +344,7 @@ function getPureArraySetter( type ) {
 
 function SingleUniform( id, activeInfo, addr ) {
 
+	this.uuid = _Math.generateUUID();
 	this.id = id;
 	this.addr = addr;
 	this.setValue = getSingularSetter( activeInfo.type );
@@ -353,6 +355,7 @@ function SingleUniform( id, activeInfo, addr ) {
 
 function PureArrayUniform( id, activeInfo, addr ) {
 
+	this.uuid = _Math.generateUUID();
 	this.id = id;
 	this.addr = addr;
 	this.size = activeInfo.size;
@@ -364,6 +367,7 @@ function PureArrayUniform( id, activeInfo, addr ) {
 
 function StructuredUniform( id ) {
 
+	this.uuid = _Math.generateUUID();
 	this.id = id;
 
 	UniformContainer.call( this ); // mix-in