When I first started using three.js I read that it supported instancing. It wasn't until I'd been using it for a while that I realised that you were supposed to use `InstancedBufferGeometry` to enable instancing. However, the issue with that is that your instances aren't actually part of the normal scene work-flow, it's pretty complicated for most users.

This on the other hand is turned on by default. Geometry will be batched and submitted to the GPU in one go as long as a browser supports the necessary extensions. An "instancing texture" is used (world and normal matrices encoded) when floating point vertex-shader accessible textures are available. On systems that don't support this, then the number of available uniforms will be detected and the necessary matrices will be pushed as arrays.

I've tried to be careful not to break backwards compatibility. The only slight issue will be for those whom are using their own shaders. They'll want to add:

```
#include <common_vertex>
```

to the top their vertex shader `main` function. This has been done already for all built-in shaders.

**EDIT**: ... _unless they're using dynamic uniforms (object render callbacks), in which case the implementation will detect that and won't turn on instancing for that object, so the user will be unaffected._

Instancing is turned on by default for opaque objects, but off for transparent ones. This is because in-order to effectively batch geometry you don't want to worry about z-sorting. Batching can be enabled for transparent objects if desired.

Performance can be tweaked by increasing the instancing texture size (`setInstancingTextureSize( size )`. By default it's 64x64, which gives you batches of size 585. Bumping it to 128x128 will use ever so slightly more VRAM, but will allow for 2340 instances per batch.

The increased VRAM is pretty negligible, but pushing a larger texture can itself be less performant than a smaller one if most of your geometry only has a few instances. For this reason, there's `minInstancingBatchSize` which will cause geometry to be rendered without batching (no texture pushing) if there's less than `minInstancingBatchSize` instances viewable in a given frame.

These properties can be changed safely between rendering frames if desired.

Sensible defaults have been chosen.

Of course, instancing can be disabled altogether by setting `autoInstancing` to `false`.