@Benjamin-Dobell 2016-11-10
As I mentioned in https://github.com/mrdoob/three.js/pull/9738, if an object has callbacks associated with it then we don't try instance those objects. This is because we have to assume the worst i.e. that those callbacks alter uniforms, in which case the objects can't be instanced.