@Benjamin-Dobell 2016-11-10
Oh, `drawMode` has been moved from `Object` to `Geometry` for the same reason... it didn't really make too much sense per object anyway.