@pailhead 2016-11-10
What are the benefits of using a texture and doing a texture fetch for each instance over using an attribute?