@pailhead 2016-11-10
I think here the uniform should be computed as an inverse float, so to avoid the division?    

uniform int instancingTextureSize;

    mat4 getModelMatrix() {

        //float dx = 1.0 / float( instancingTextureSize );
        //float dy = 1.0 / float( instancingTextureSize );
        float dx =  instancingTexture_INVERSE_Size ;
        float dy =  instancingTexture_INVERSE_Size ;