@Benjamin-Dobell 2016-11-10
**Note**: Force push just modified comments, nothing functional.

@pailhead Yeah, I didn't try optimise it too much. I originally planned to mimic the skinning shader (bone texture), but I couldn't help myself optimising it at least a little bit ;)

@pailhead Attributes are typically designed to be one-per-vertex. Of course the ANGLE instance rendering functionality lets you set a divisor. I've kind of assumed that texture lookups will be more performant, but that may well be complete nonsense. There's both textures and uniform options, I guess there's nothing stopping someone adding an attributes option as well.

_I actually implemented something very similar, albeit more advanced, to this years ago for Xbox 360, attributes wasn't an option then. But I guess I started in a "texture mindset" for that reason. However, the plan for the texture was that it actually that in the future it could be extended for fully instanced, unlimited bone, skinned models (which is what I'd implemented in the past)._