@pailhead 2016-11-10
rather than textures->uniforms->attributes i think it should go in reverse :)