@Benjamin-Dobell 2016-11-10
Well, textures makes sense of uniforms due to constraints on how many uniforms you're going to have available. On my system you'd get about 140 instances per batch vs. 585.

However, you're welcome to mess around with this and give attributes a shot. If you don't beat me to it, then I'll take a crack at it when I've got a bit more time.