@Benjamin-Dobell 2016-11-10
Oh yeah, the other limitation with attributes is how many you can push per-vertex (typically 8-16). We'd need 7 (6 with clever packing), however that's possibly going to interfere with custom shaders/materials etc. (if it works at all on systems with lower limits). But yeah, that certainly doesn't totally preclude the idea of using attributes.