@pailhead 2016-11-10
I didn't have a chance to go through the entire code, i don't know how it figures out what to batch, but it would definitely be worth doing some benchmarks. 
I tried to approach this slightly differently, but im on an older version of three. I basically made a class that creates an instanced buffer geometry out of a regular buffer geometry and a "positioning" callback. The regular geometry populates the mesh attributes, while the positioning callback runs some logic and fills the instancing attributes. 

This is a different approach though, i have to know ahead of time what im instancing and how im doing it. 

I'm a bit confused with where three.js stands now. It's not really a game engine so i'm not sure if it should be handling optimizations like this. IE. first we'd have to make sure that saving a few draw calls, is worth having each mesh now do 7 texture fetches. 

On the other hand if you know that you want to draw something in a specific way, you'd set this up manually, ie. by making something with `instancedBufferGeometry` in the first place, and populating it's attributes based on some logic, or even choosing a texture or a uniform over that. 


For my use case i didn't actually hack much, the convenience class to obtain the instancing data is something you can build with existing classes. The only thing i did have to hack in was the TRS matrix. The assumption made there was that at a bare minimum, something needs to have position,rotation,scale, and this should be exposed to all the shaders. I basically added a ".instanceTransform" property to the base material class which triggers a define in the shader.