@pailhead 2016-11-10
> We'd need 7 (6 with clever packing),

Are you 100% on this? I think you only need TRS for this, not the normal matrix. I need to look at my code and if something has changed, but i'm pretty sure that you can transform the normals just once, in object space.