@Benjamin-Dobell 2016-11-10
The biggest issue for me is that `InstancedGeometryBuffer` circumvents the scene graph, for me that's a deal breaker.

Also I don't really see why anyone should be expected to handle this manually when it can trivially be performed for them automatically.

I've benchmarked this for my own needs. My geometry was already heavily optimised for mobile, but on average I'm seeing frame rendering time reduced by 25%. Gains will be proportionally much larger if you're not auto updating transforms (with is presently responsible for around 50% of my render loop). Also if you're using larger pieces of geometry you'll have bigger gains. My scenes are unfortunately heavily broken up as it aided collision detection.