@Benjamin-Dobell 2016-11-10
Normal matrix requires an invert operation. Not included in the current WebGL GLSL spec and extremely expensive to implement manually.