@Benjamin-Dobell 2016-11-10
Oh, yeah you don't *need* a normal matrix, I didn't use one when I implemented this last time. But having one makes for simpler calculations in the shader. Also I wanted these changes to be as transparent as possible I.e. Not rewriting all the lighting shaders.