@pailhead 2016-11-10
what do you mean by circumnavigating the scene graph?