@pailhead 2016-11-10
> Oh, yeah you don't need a normal matrix, 

I didn't mean it like that, i forgot that it's just the rotation extracted, i mostly render real world objects so they're only uniformly scaled :)

Was thinking about the question one might ask before implementing something like this. Like "do my objects move, or are they static" "will i add more of them" "does the geometry need to be queried in any way" etc etc. 

InstancedBufferGeometry shouldn't have anything to do with the scenegraph in the first place, pretty much the same way BufferGeometry shouldn't, but i may be missing something?