@Benjamin-Dobell 2016-11-10
Well, this implementation is seamless, it shouldn't impact a single example, if it does it's a bug.

With InstancedBufferGeometry you only add one object to the scene, the instances are defined by some other means (custom shader code, attributes etc.)

This implementation just goes "Well, several objects already share one Geometry. Might as well share a draw call too." It's not really any more advanced than that. For particle effects etc. you'd still use something like InstancedBufferGeometry.