@pailhead 2016-11-10
I think the question/conclusion doesn't stop there - "For every instance of this geometry drawn, i will do 7 texture fetches for each of it's vertices". I think this shouldn't really be hardcoded in the library. 

> With InstancedBufferGeometry you only add one object to the scene,

What's preventing you from adding other objects to the scene? You can make an arbitrary amount of meshes point to the instanceBufferGeometry, or even keep an instance of BufferGeometry referring to the same geometry attributes. If you compute something like bounding box or bounding sphere, you can copy those to other objects too. There really shouldn't be a conceptual difference between these two, it just branches the WebGLRenderer and does slightly different calls. There is no reason why you would only use one object when using either of the geometries.