@Benjamin-Dobell 2017-02-03
Alright. Here are the latest updates:

- Rebased on dev once more (as there were conflicts again).
- Fixed a MultiMaterial bug that was causing non-indexed buffer geometry to be rendered entirely for each each material (rather than just rendering the group corresponding to the sub-materials). Visual fix + improved performance (9cdc2ff)
- Increased the default autoInstancingMinBatchSize from 2 to 20. (	b3612a2)
  This was just based on observations on my own machine that bufferSubData (https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/bufferSubData) does not perform all that well. So it's not worth instancing small groups of the one object. NOTE: Small groups of the one object with huge amounts of vertices would benefit, however I don't think really high-poly models is the normal use case for three.js (and of course, this is just the default setting and can be tweaked as desired).
- Fixed a bug where auto instancing is enabled, but rendering pieces of geometry <  autoInstancingMinBatchSize incorrectly caused their material to be marked as `needsUpdate` (	b07dc57).
- Formatting fixes (indentation, whitespace placement etc.).
- Prevented performance degradation if the geometry-material paris being rendered changes significantly / often. (1834eb1)


@lojjic That last commit was the main cause of slow-down with respect to your test, as each time you toggled the tick-box thousands of different items were being rendered. Toggling it over and over work cause performance to get worse and worse. That's no longer the case.

Test results on my computer are now:

## This pull request

### Instanceable mesh (3 materials)

Attributes Instancing: 19ms / frame
Texture Instancing: 19ms / frame
Instancing Disabled: 67ms / frame

### Non-instanceable mesh (8000 materials)

Attributes Instancing: 100ms / frame
Texture Instancing: 100ms / frame
Instancing Disabled: 80ms / frame

## Current dev head (da9f5a68e48bd855b1acf4a4ea1076e69a8a477a)

### Instanceable mesh (3 materials)

61ms / frame

### Non-instanceable mesh (8000 materials)

83ms / frame

### Explanation for performance difference

Well, when instanceable it's obvious why this pull request is faster. However, when there are 8000 materials on-screen this pull request has worse performance than the current dev head.

This simply comes down to the fact that when instancing we have to group up the scene objects (mesh) into geometry-material pairs. We basically end up with a hash with 8000 entries, and each entry contains only 1 instance. If auto instancing is disabled on this branch there's still a slight performance hit in this particular example, but it's pretty negligible considering the setup.

I'm pretty stoked with these results; as the branch is still doing quite well under somewhat contrived condition. It's a great test to determine the systems limits, but probably not particularly accurate with respect to real world usage. I'd expect that it's _extremely_ unlikely that users have 8000 different materials in the one scene, certainly not visible in the one frame.

Typically only _really_ huge scenes would have anywhere near that many materials. However, huge scenes would also typically have lots of instanceable geometry - which is the kind of thing this branch would excel at.

@lojjic What do you think?